==> Chapter-4/04_descriptives.Rmd <==
---
title: "2.1_descriptives"
---

```{r setup}
# Load packages
library("tidyverse")
library("survminer")
library("survival")
library("pammtools")
library("openxlsx")
library("psfmi")
library("gtsummary")

rm(list = ls())

```

# Sample statistics
```{r}
rm(list = ls())

load(".../data_final_uk.RData")

load(".../data_final_de.RData")

data_uk_imp_1$country <- "United Kingdom"

data_de_imp_1$country <- "Germany"

data_combined <- rbind(data_uk_imp_1, data_de_imp_1)

rm(list = setdiff(ls(), c("data_combined")))

results <- survey::svydesign(~ 1, data = data_combined, weights = ~ final_weight) %>%
  tbl_svysummary(
    by = country,
    include = c(age, retired, homeowner_type, homeowner_type_type, social_housing, female, year_born, migrant, education_level, poor_health, marital_status, children_any, household_size, occupation_group, public_sector, self_employed, hh_income_eq, unemployment_rate, inflation_rate, region),
    statistic = list(all_categorical() ~ "{p}", all_continuous() ~ "{mean} ({sd})"),
    digits = ~ 1)

results

```

# Missingness
```{r}
# Clear workspace
rm(list = ls())

# Load results files
load(".../data_final_uk.RData")

data <- data_de %>% select(-c("country", "orgpid", "age", "female", "year_born"))

# Check overall missingness
(sum(is.na(data)) * 100 )/(prod(dim(data)))

# Check missingness in each variable
round(colSums(is.na(data))/nrow(data) * 100, 1)

# Check observations with at least one missing
round(100 - (data %>% na.omit() %>% nrow()) / nrow(data) * 100, 1)

```

# Survival plots
```{r}
rm(list = ls())

load(".../data_final_uk.RData")

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_1 %>% mutate(age1 = age - 1), weights = final_weight)
surv_1 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_2 %>% mutate(age1 = age - 1), weights = final_weight)
surv_2 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_3 %>% mutate(age1 = age - 1), weights = final_weight)
surv_3 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_4 %>% mutate(age1 = age - 1), weights = final_weight)
surv_4 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_5 %>% mutate(age1 = age - 1), weights = final_weight)
surv_5 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_6 %>% mutate(age1 = age - 1), weights = final_weight)
surv_6 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_7 %>% mutate(age1 = age - 1), weights = final_weight)
surv_7 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_8 %>% mutate(age1 = age - 1), weights = final_weight)
surv_8 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_9 %>% mutate(age1 = age - 1), weights = final_weight)
surv_9 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_10 %>% mutate(age1 = age - 1), weights = final_weight)
surv_10 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_11 %>% mutate(age1 = age - 1), weights = final_weight)
surv_11 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_12 %>% mutate(age1 = age - 1), weights = final_weight)
surv_12 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_13 %>% mutate(age1 = age - 1), weights = final_weight)
surv_13 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_14 %>% mutate(age1 = age - 1), weights = final_weight)
surv_14 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_15 %>% mutate(age1 = age - 1), weights = final_weight)
surv_15 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_16 %>% mutate(age1 = age - 1), weights = final_weight)
surv_16 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_17 %>% mutate(age1 = age - 1), weights = final_weight)
surv_17 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_18 %>% mutate(age1 = age - 1), weights = final_weight)
surv_18 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_19 %>% mutate(age1 = age - 1), weights = final_weight)
surv_19 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_20 %>% mutate(age1 = age - 1), weights = final_weight)
surv_20 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

survival_combined_uk <- rbind(surv_1, surv_2, surv_3, surv_4, surv_5, surv_6, surv_7, surv_8, surv_9, surv_10)

survival_combined_uk <- survival_combined_uk %>%
  group_by(homeowner_type, time) %>% mutate(
    survival = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[3]) %>%
  select(homeowner_type, time, survival, conf_low, conf_high) %>% unique()

survival_combined_uk <- survival_combined_uk %>%
  group_by(homeowner_type) %>%
  complete(time = 50:66) %>%
  arrange(homeowner_type, time) %>%
  fill(1:5, .direction = "downup") %>% unique()

load(".../data_final_de.RData")

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_1 %>% mutate(age1 = age - 1), weights = final_weight)
surv_1 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_2 %>% mutate(age1 = age - 1), weights = final_weight)
surv_2 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_3 %>% mutate(age1 = age - 1), weights = final_weight)
surv_3 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_4 %>% mutate(age1 = age - 1), weights = final_weight)
surv_4 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_5 %>% mutate(age1 = age - 1), weights = final_weight)
surv_5 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_6 %>% mutate(age1 = age - 1), weights = final_weight)
surv_6 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_7 %>% mutate(age1 = age - 1), weights = final_weight)
surv_7 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_8 %>% mutate(age1 = age - 1), weights = final_weight)
surv_8 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_9 %>% mutate(age1 = age - 1), weights = final_weight)
surv_9 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_10 %>% mutate(age1 = age - 1), weights = final_weight)
surv_10 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_11 %>% mutate(age1 = age - 1), weights = final_weight)
surv_11 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_12 %>% mutate(age1 = age - 1), weights = final_weight)
surv_12 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_13 %>% mutate(age1 = age - 1), weights = final_weight)
surv_13 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_14 %>% mutate(age1 = age - 1), weights = final_weight)
surv_14 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_15 %>% mutate(age1 = age - 1), weights = final_weight)
surv_15 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_16 %>% mutate(age1 = age - 1), weights = final_weight)
surv_16 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_17 %>% mutate(age1 = age - 1), weights = final_weight)
surv_17 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_18 %>% mutate(age1 = age - 1), weights = final_weight)
surv_18 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_19 %>% mutate(age1 = age - 1), weights = final_weight)
surv_19 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_de_imp_20 %>% mutate(age1 = age - 1), weights = final_weight)
surv_20 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

survival_combined_de <- rbind(surv_1, surv_2, surv_3, surv_4, surv_5, surv_6, surv_7, surv_8, surv_9, surv_10, surv_11, surv_12, surv_13, surv_14, surv_15, surv_16, surv_17, surv_18, surv_19, surv_20)

survival_combined_de <- survival_combined_de %>%
  group_by(homeowner_type, time) %>% mutate(
    survival = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[3]) %>%
  select(homeowner_type, time, survival, conf_low, conf_high) %>% unique()

survival_combined_de <- survival_combined_de %>%
  group_by(homeowner_type) %>%
  complete(time = 50:66) %>%
  arrange(homeowner_type, time) %>%
  fill(1:5, .direction = "downup") %>% unique()

survival_uk <- survival_combined_uk %>% mutate(country = "United Kingdom")

survival_de <- survival_combined_de %>% mutate(country = "Germany")

survival <- rbind(survival_uk, survival_de)

survival <- survival %>% 
  mutate(homeowner_type = case_when(
    homeowner_type == "Outright owner" ~ "Outright owners",
    homeowner_type == "Mortgagor" ~ "Owners with mortgage",
    homeowner_type == "Renter" ~ "Renters"))

surv_plot <- ggplot(survival, aes(x = time - 0.5, y = 1 - survival, group = homeowner_type, fill = homeowner_type)) +
  facet_grid(. ~ country) +
  geom_stepribbon(aes(ymin = 1 - conf_low, ymax = 1 - conf_high), alpha = 0.25) +
  geom_step(aes(color = homeowner_type), size = 0.8) +
  scale_x_continuous("Age", limits = c(50, 66), breaks = seq(50, 65, 1)) +
  scale_y_continuous("Cumulative risk (% retired)", breaks = seq(0, 1, 0.1), label = scales::percent) +
  scale_fill_manual(values = c("darkblue", "darkgreen", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkgreen", "darkred")) +
  theme_linedraw() +
  coord_cartesian(ylim = c(0,1), xlim = c(51, 65)) +
  theme(
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.y = element_text(size = 13, face = "bold", colour = "black"), 
        axis.title.x = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_blank(),
        plot.background = element_blank(),
        strip.text.x = element_text(size = 13, colour = "white"),
        legend.position=c(.17,.91),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.text = element_text(colour = "black", size = 12),
        legend.background = element_rect(fill = 'transparent'),
        legend.title = element_blank())
ggsave("final_plots/plot_survival_type.pdf", height = 5.5, width = 7.5, bg = "transparent", surv_plot)

```

```{r}

```
