==> Chapter-4/05_results_plots.Rmd <==
---
author: "Jan Einhoff"
date: "06/2025"
---

This R markdown file is used to create the cumulative risk and risk difference plots.

```{r setup}
# Load packages
library("tidyverse")
library("psfmi")
library("pammtools")
library("gridExtra")

# Clear work space
rm(list = ls())

options("scipen" = 100, "digits" = 4)

```

```{r}
load("/Users/jlein/Documents/analysis_homeowner/data/data_final_uk.RData")

data <- data_uk

rm(list = setdiff(ls(), "data"))

# Load data
load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp1.RData")
results_imp1 <- results %>% mutate(imp = 1) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp1_r_h <- contrasts_renters_owners %>% mutate(imp = 1)
contrasts_imp1_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 1)
contrasts_imp1_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 1)
contrasts_imp1_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 1)
contrasts_imp1_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 1)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp2.RData")
results_imp2 <- results %>% mutate(imp = 2) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp2_r_h <- contrasts_renters_owners %>% mutate(imp = 2)
contrasts_imp2_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 2)
contrasts_imp2_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 2)
contrasts_imp2_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 2)
contrasts_imp2_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 2)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp3.RData")
results_imp3 <- results %>% mutate(imp = 3) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp3_r_h <- contrasts_renters_owners %>% mutate(imp = 3)
contrasts_imp3_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 3)
contrasts_imp3_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 3)
contrasts_imp3_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 3)
contrasts_imp3_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 3)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp4.RData")
results_imp4 <- results %>% mutate(imp = 4) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp4_r_h <- contrasts_renters_owners %>% mutate(imp = 4)
contrasts_imp4_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 4)
contrasts_imp4_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 4)
contrasts_imp4_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 4)
contrasts_imp4_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 4)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp5.RData")
results_imp5 <- results %>% mutate(imp = 5) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp5_r_h <- contrasts_renters_owners %>% mutate(imp = 5)
contrasts_imp5_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 5)
contrasts_imp5_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 5)
contrasts_imp5_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 5)
contrasts_imp5_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 5)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp6.RData")
results_imp6 <- results %>% mutate(imp = 6) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp6_r_h <- contrasts_renters_owners %>% mutate(imp = 6)
contrasts_imp6_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 6)
contrasts_imp6_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 6)
contrasts_imp6_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 6)
contrasts_imp6_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 6)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp7.RData")
results_imp7 <- results %>% mutate(imp = 7) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp7_r_h <- contrasts_renters_owners %>% mutate(imp = 7)
contrasts_imp7_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 7)
contrasts_imp7_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 7)
contrasts_imp7_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 7)
contrasts_imp7_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 7)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp8.RData")
results_imp8 <- results %>% mutate(imp = 8) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp8_r_h <- contrasts_renters_owners %>% mutate(imp = 8)
contrasts_imp8_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 8)
contrasts_imp8_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 8)
contrasts_imp8_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 8)
contrasts_imp8_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 8)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp9.RData")
results_imp9 <- results %>% mutate(imp = 9) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp9_r_h <- contrasts_renters_owners %>% mutate(imp = 9)
contrasts_imp9_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 9)
contrasts_imp9_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 9)
contrasts_imp9_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 9)
contrasts_imp9_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 9)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp10.RData")
results_imp10 <- results %>% mutate(imp = 10) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp10_r_h <- contrasts_renters_owners %>% mutate(imp = 10)
contrasts_imp10_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 10)
contrasts_imp10_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 10)
contrasts_imp10_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 10)
contrasts_imp10_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 10)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp11.RData")
results_imp11 <- results %>% mutate(imp = 11) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp11_r_h <- contrasts_renters_owners %>% mutate(imp = 11)
contrasts_imp11_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 11)
contrasts_imp11_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 11)
contrasts_imp11_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 11)
contrasts_imp11_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 11)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp12.RData")
results_imp12 <- results %>% mutate(imp = 12) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp12_r_h <- contrasts_renters_owners %>% mutate(imp = 12)
contrasts_imp12_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 12)
contrasts_imp12_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 12)
contrasts_imp12_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 12)
contrasts_imp12_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 12)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp13.RData")
results_imp13 <- results %>% mutate(imp = 13) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp13_r_h <- contrasts_renters_owners %>% mutate(imp = 13)
contrasts_imp13_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 13)
contrasts_imp13_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 13)
contrasts_imp13_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 13)
contrasts_imp13_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 13)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp14.RData")
results_imp14 <- results %>% mutate(imp = 14) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp14_r_h <- contrasts_renters_owners %>% mutate(imp = 14)
contrasts_imp14_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 14)
contrasts_imp14_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 14)
contrasts_imp14_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 14)
contrasts_imp14_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 14)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp15.RData")
results_imp15 <- results %>% mutate(imp = 15) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp15_r_h <- contrasts_renters_owners %>% mutate(imp = 15)
contrasts_imp15_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 15)
contrasts_imp15_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 15)
contrasts_imp15_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 15)
contrasts_imp15_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 15)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp16.RData")
results_imp16 <- results %>% mutate(imp = 16) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp16_r_h <- contrasts_renters_owners %>% mutate(imp = 16)
contrasts_imp16_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 16)
contrasts_imp16_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 16)
contrasts_imp16_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 16)
contrasts_imp16_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 16)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp17.RData")
results_imp17 <- results %>% mutate(imp = 17) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp17_r_h <- contrasts_renters_owners %>% mutate(imp = 17)
contrasts_imp17_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 17)
contrasts_imp17_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 17)
contrasts_imp17_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 17)
contrasts_imp17_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 17)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp18.RData")
results_imp18 <- results %>% mutate(imp = 18) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp18_r_h <- contrasts_renters_owners %>% mutate(imp = 18)
contrasts_imp18_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 18)
contrasts_imp18_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 18)
contrasts_imp18_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 18)
contrasts_imp18_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 18)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp19.RData")
results_imp19 <- results %>% mutate(imp = 19) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp19_r_h <- contrasts_renters_owners %>% mutate(imp = 19)
contrasts_imp19_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 19)
contrasts_imp19_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 19)
contrasts_imp19_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 19)
contrasts_imp19_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 19)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_uk_975_imp20.RData")
results_imp20 <- results %>% mutate(imp = 20) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp20_r_h <- contrasts_renters_owners %>% mutate(imp = 20)
contrasts_imp20_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 20)
contrasts_imp20_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 20)
contrasts_imp20_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 20)
contrasts_imp20_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 20)

survival_combined <- rbind(results_imp1, results_imp2, results_imp3, results_imp4, results_imp5,
                           results_imp6, results_imp7, results_imp8, results_imp9, results_imp10,
                           results_imp11, results_imp12, results_imp13, results_imp14, results_imp15,
                           results_imp16, results_imp17, results_imp18, results_imp19, results_imp20)

contrast_combined_r_h <- rbind(
  contrasts_imp1_r_h, contrasts_imp2_r_h, contrasts_imp3_r_h, contrasts_imp4_r_h, contrasts_imp5_r_h,
  contrasts_imp6_r_h, contrasts_imp7_r_h, contrasts_imp8_r_h, contrasts_imp9_r_h, contrasts_imp10_r_h,
  contrasts_imp11_r_h, contrasts_imp12_r_h, contrasts_imp13_r_h, contrasts_imp14_r_h, contrasts_imp15_r_h,
  contrasts_imp16_r_h, contrasts_imp17_r_h, contrasts_imp18_r_h, contrasts_imp19_r_h, contrasts_imp20_r_h)

contrast_combined_r_m <- rbind(
  contrasts_imp1_r_m, contrasts_imp2_r_m, contrasts_imp3_r_m, contrasts_imp4_r_m, contrasts_imp5_r_m,
  contrasts_imp6_r_m, contrasts_imp7_r_m, contrasts_imp8_r_m, contrasts_imp9_r_m, contrasts_imp10_r_m,
  contrasts_imp11_r_m, contrasts_imp12_r_m, contrasts_imp13_r_m, contrasts_imp14_r_m, contrasts_imp15_r_m,
  contrasts_imp16_r_m, contrasts_imp17_r_m, contrasts_imp18_r_m, contrasts_imp19_r_m, contrasts_imp20_r_m)

contrast_combined_r_o <- rbind(
  contrasts_imp1_r_o, contrasts_imp2_r_o, contrasts_imp3_r_o, contrasts_imp4_r_o, contrasts_imp5_r_o,
  contrasts_imp6_r_o, contrasts_imp7_r_o, contrasts_imp8_r_o, contrasts_imp9_r_o, contrasts_imp10_r_o,
  contrasts_imp11_r_o, contrasts_imp12_r_o, contrasts_imp13_r_o, contrasts_imp14_r_o, contrasts_imp15_r_o,
  contrasts_imp16_r_o, contrasts_imp17_r_o, contrasts_imp18_r_o, contrasts_imp19_r_o, contrasts_imp20_r_o)

contrast_combined_o_m <- rbind(
  contrasts_imp1_o_m, contrasts_imp2_o_m, contrasts_imp3_o_m, contrasts_imp4_o_m, contrasts_imp5_o_m,
  contrasts_imp6_o_m, contrasts_imp7_o_m, contrasts_imp8_o_m, contrasts_imp9_o_m, contrasts_imp10_o_m,
  contrasts_imp11_o_m, contrasts_imp12_o_m, contrasts_imp13_o_m, contrasts_imp14_o_m, contrasts_imp15_o_m,
  contrasts_imp16_o_m, contrasts_imp17_o_m, contrasts_imp18_o_m, contrasts_imp19_o_m, contrasts_imp20_o_m)

contrast_combined_rp_o <- rbind(
  contrasts_imp1_rp_o, contrasts_imp2_rp_o, contrasts_imp3_rp_o, contrasts_imp4_rp_o, contrasts_imp5_rp_o,
  contrasts_imp6_rp_o, contrasts_imp7_rp_o, contrasts_imp8_rp_o, contrasts_imp9_rp_o, contrasts_imp10_rp_o,
  contrasts_imp11_rp_o, contrasts_imp12_rp_o, contrasts_imp13_rp_o, contrasts_imp14_rp_o, contrasts_imp15_rp_o,
  contrasts_imp16_rp_o, contrasts_imp17_rp_o, contrasts_imp18_rp_o, contrasts_imp19_rp_o, contrasts_imp20_rp_o)

```

```{r}
survival_uk <- survival_combined %>%
  group_by(group, age) %>%
  summarise(
    survival = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  group_by(group) %>%
  complete(age = 50:66) %>%
  arrange(group, age) %>%
  fill(1:6, .direction = "downup") %>%
  ungroup()

contrast_uk_r_h <- contrast_combined_r_h %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_uk_r_m <- contrast_combined_r_m %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_uk_r_o <- contrast_combined_r_o %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_uk_o_m <- contrast_combined_o_m %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_uk_rp_o <- contrast_combined_rp_o %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

survival_uk_r_h <- survival_uk %>% filter(group %in% c("Owners", "Renters")) %>% mutate(country = "United Kingdom")
survival_uk_r_h <- survival_uk_r_h %>% 
  mutate(group = case_when(
    group == "Owners" ~ "Owning",
    group == "Renters" ~ "Renting"))

survival_uk_r_m_o <- survival_uk %>% filter(group %in% c("Owners outright", "Owners mortgage", "Renters")) %>% mutate(country = "United Kingdom")
survival_uk_r_m_o <- survival_uk_r_m_o %>% 
  mutate(group = case_when(
    group == "Owners mortgage" ~ "Owning with mortgage",
    group == "Owners outright" ~ "Owning outright",
    group == "Renters" ~ "Renting"))

survival_uk_rp_o <- survival_uk %>% filter(group %in% c("Owners", "Renters (private)")) %>% mutate(country = "United Kingdom")
survival_uk_rp_o <- survival_uk_rp_o %>% 
  mutate(group = case_when(
    group == "Owners" ~ "Owning",
    group == "Renters (private)" ~ "Renting (private)"))

contrast_uk_r_h <- contrast_uk_r_h %>% mutate(country = "United Kingdom", contrast = "Owning")

contrast_uk_r_m <- contrast_uk_r_m %>% mutate(country = "United Kingdom", contrast = "Owning with mortgage")

contrast_uk_r_o <- contrast_uk_r_o %>% mutate(country = "United Kingdom", contrast = "Owning outright")

contrast_uk_r_m_o <- rbind(contrast_uk_r_m, contrast_uk_r_o)

contrast_uk_o_m <- contrast_uk_o_m %>% mutate(country = "United Kingdom", contrast = "Owning outright vs. \nOwning with mortgage")

contrast_uk_rp_o <- contrast_uk_rp_o %>% mutate(country = "United Kingdom", contrast = "Private renting vs. Owning")

```

```{r}
load("/Users/jlein/Documents/analysis_homeowner/data/data_final_de.RData")

data <- data_de

# Load data
data <- data_de

# Load data
load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp1.RData")
results_imp1 <- results %>% mutate(imp = 1) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp1_r_h <- contrasts_renters_owners %>% mutate(imp = 1)
contrasts_imp1_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 1)
contrasts_imp1_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 1)
contrasts_imp1_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 1)
contrasts_imp1_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 1)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp2.RData")
results_imp2 <- results %>% mutate(imp = 2) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp2_r_h <- contrasts_renters_owners %>% mutate(imp = 2)
contrasts_imp2_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 2)
contrasts_imp2_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 2)
contrasts_imp2_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 2)
contrasts_imp2_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 2)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp3.RData")
results_imp3 <- results %>% mutate(imp = 3) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp3_r_h <- contrasts_renters_owners %>% mutate(imp = 3)
contrasts_imp3_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 3)
contrasts_imp3_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 3)
contrasts_imp3_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 3)
contrasts_imp3_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 3)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp4.RData")
results_imp4 <- results %>% mutate(imp = 4) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp4_r_h <- contrasts_renters_owners %>% mutate(imp = 4)
contrasts_imp4_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 4)
contrasts_imp4_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 4)
contrasts_imp4_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 4)
contrasts_imp4_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 4)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp5.RData")
results_imp5 <- results %>% mutate(imp = 5) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp5_r_h <- contrasts_renters_owners %>% mutate(imp = 5)
contrasts_imp5_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 5)
contrasts_imp5_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 5)
contrasts_imp5_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 5)
contrasts_imp5_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 5)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp6.RData")
results_imp6 <- results %>% mutate(imp = 6) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp6_r_h <- contrasts_renters_owners %>% mutate(imp = 6)
contrasts_imp6_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 6)
contrasts_imp6_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 6)
contrasts_imp6_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 6)
contrasts_imp6_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 6)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp7.RData")
results_imp7 <- results %>% mutate(imp = 7) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp7_r_h <- contrasts_renters_owners %>% mutate(imp = 7)
contrasts_imp7_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 7)
contrasts_imp7_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 7)
contrasts_imp7_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 7)
contrasts_imp7_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 7)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp8.RData")
results_imp8 <- results %>% mutate(imp = 8) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp8_r_h <- contrasts_renters_owners %>% mutate(imp = 8)
contrasts_imp8_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 8)
contrasts_imp8_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 8)
contrasts_imp8_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 8)
contrasts_imp8_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 8)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp9.RData")
results_imp9 <- results %>% mutate(imp = 9) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp9_r_h <- contrasts_renters_owners %>% mutate(imp = 9)
contrasts_imp9_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 9)
contrasts_imp9_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 9)
contrasts_imp9_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 9)
contrasts_imp9_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 9)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp10.RData")
results_imp10 <- results %>% mutate(imp = 10) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp10_r_h <- contrasts_renters_owners %>% mutate(imp = 10)
contrasts_imp10_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 10)
contrasts_imp10_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 10)
contrasts_imp10_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 10)
contrasts_imp10_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 10)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp11.RData")
results_imp11 <- results %>% mutate(imp = 11) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp11_r_h <- contrasts_renters_owners %>% mutate(imp = 11)
contrasts_imp11_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 11)
contrasts_imp11_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 11)
contrasts_imp11_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 11)
contrasts_imp11_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 11)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp12.RData")
results_imp12 <- results %>% mutate(imp = 12) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp12_r_h <- contrasts_renters_owners %>% mutate(imp = 12)
contrasts_imp12_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 12)
contrasts_imp12_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 12)
contrasts_imp12_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 12)
contrasts_imp12_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 12)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp13.RData")
results_imp13 <- results %>% mutate(imp = 13) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp13_r_h <- contrasts_renters_owners %>% mutate(imp = 13)
contrasts_imp13_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 13)
contrasts_imp13_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 13)
contrasts_imp13_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 13)
contrasts_imp13_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 13)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp14.RData")
results_imp14 <- results %>% mutate(imp = 14) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp14_r_h <- contrasts_renters_owners %>% mutate(imp = 14)
contrasts_imp14_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 14)
contrasts_imp14_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 14)
contrasts_imp14_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 14)
contrasts_imp14_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 14)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp15.RData")
results_imp15 <- results %>% mutate(imp = 15) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp15_r_h <- contrasts_renters_owners %>% mutate(imp = 15)
contrasts_imp15_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 15)
contrasts_imp15_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 15)
contrasts_imp15_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 15)
contrasts_imp15_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 15)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp16.RData")
results_imp16 <- results %>% mutate(imp = 16) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp16_r_h <- contrasts_renters_owners %>% mutate(imp = 16)
contrasts_imp16_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 16)
contrasts_imp16_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 16)
contrasts_imp16_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 16)
contrasts_imp16_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 16)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp17.RData")
results_imp17 <- results %>% mutate(imp = 17) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp17_r_h <- contrasts_renters_owners %>% mutate(imp = 17)
contrasts_imp17_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 17)
contrasts_imp17_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 17)
contrasts_imp17_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 17)
contrasts_imp17_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 17)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp18.RData")
results_imp18 <- results %>% mutate(imp = 18) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp18_r_h <- contrasts_renters_owners %>% mutate(imp = 18)
contrasts_imp18_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 18)
contrasts_imp18_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 18)
contrasts_imp18_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 18)
contrasts_imp18_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 18)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp19.RData")
results_imp19 <- results %>% mutate(imp = 19) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp19_r_h <- contrasts_renters_owners %>% mutate(imp = 19)
contrasts_imp19_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 19)
contrasts_imp19_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 19)
contrasts_imp19_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 19)
contrasts_imp19_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 19)

load("/Users/jlein/Documents/analysis_homeowner/final_results/output_de_975_imp20.RData")
results_imp20 <- results %>% mutate(imp = 20) %>% group_by(age) %>% mutate(group = c("Renters", "Renters (private)", "Owners", "Owners outright", "Owners mortgage"))
contrasts_imp20_r_h <- contrasts_renters_owners %>% mutate(imp = 20)
contrasts_imp20_r_m <- contrasts_renters_owners_mortgage %>% mutate(imp = 20)
contrasts_imp20_r_o <- contrasts_renters_owners_outright %>% mutate(imp = 20)
contrasts_imp20_o_m <- contrasts_owners_outright_owners_mortgage %>% mutate(imp = 20)
contrasts_imp20_rp_o <- contrasts_renters_private_owners %>% mutate(imp = 20)

survival_combined <- rbind(results_imp1, results_imp2, results_imp3, results_imp4, results_imp5,
                           results_imp6, results_imp7, results_imp8, results_imp9, results_imp10,
                           results_imp11, results_imp12, results_imp13, results_imp14, results_imp15,
                           results_imp16, results_imp17, results_imp18, results_imp19, results_imp20)

contrast_combined_r_h <- rbind(
  contrasts_imp1_r_h, contrasts_imp2_r_h, contrasts_imp3_r_h, contrasts_imp4_r_h, contrasts_imp5_r_h,
  contrasts_imp6_r_h, contrasts_imp7_r_h, contrasts_imp8_r_h, contrasts_imp9_r_h, contrasts_imp10_r_h,
  contrasts_imp11_r_h, contrasts_imp12_r_h, contrasts_imp13_r_h, contrasts_imp14_r_h, contrasts_imp15_r_h,
  contrasts_imp16_r_h, contrasts_imp17_r_h, contrasts_imp18_r_h, contrasts_imp19_r_h, contrasts_imp20_r_h)

contrast_combined_r_m <- rbind(
  contrasts_imp1_r_m, contrasts_imp2_r_m, contrasts_imp3_r_m, contrasts_imp4_r_m, contrasts_imp5_r_m,
  contrasts_imp6_r_m, contrasts_imp7_r_m, contrasts_imp8_r_m, contrasts_imp9_r_m, contrasts_imp10_r_m,
  contrasts_imp11_r_m, contrasts_imp12_r_m, contrasts_imp13_r_m, contrasts_imp14_r_m, contrasts_imp15_r_m,
  contrasts_imp16_r_m, contrasts_imp17_r_m, contrasts_imp18_r_m, contrasts_imp19_r_m, contrasts_imp20_r_m)

contrast_combined_r_o <- rbind(
  contrasts_imp1_r_o, contrasts_imp2_r_o, contrasts_imp3_r_o, contrasts_imp4_r_o, contrasts_imp5_r_o,
  contrasts_imp6_r_o, contrasts_imp7_r_o, contrasts_imp8_r_o, contrasts_imp9_r_o, contrasts_imp10_r_o,
  contrasts_imp11_r_o, contrasts_imp12_r_o, contrasts_imp13_r_o, contrasts_imp14_r_o, contrasts_imp15_r_o,
  contrasts_imp16_r_o, contrasts_imp17_r_o, contrasts_imp18_r_o, contrasts_imp19_r_o, contrasts_imp20_r_o)

contrast_combined_o_m <- rbind(
  contrasts_imp1_o_m, contrasts_imp2_o_m, contrasts_imp3_o_m, contrasts_imp4_o_m, contrasts_imp5_o_m,
  contrasts_imp6_o_m, contrasts_imp7_o_m, contrasts_imp8_o_m, contrasts_imp9_o_m, contrasts_imp10_o_m,
  contrasts_imp11_o_m, contrasts_imp12_o_m, contrasts_imp13_o_m, contrasts_imp14_o_m, contrasts_imp15_o_m,
  contrasts_imp16_o_m, contrasts_imp17_o_m, contrasts_imp18_o_m, contrasts_imp19_o_m, contrasts_imp20_o_m)

contrast_combined_rp_o <- rbind(
  contrasts_imp1_rp_o, contrasts_imp2_rp_o, contrasts_imp3_rp_o, contrasts_imp4_rp_o, contrasts_imp5_rp_o,
  contrasts_imp6_rp_o, contrasts_imp7_rp_o, contrasts_imp8_rp_o, contrasts_imp9_rp_o, contrasts_imp10_rp_o,
  contrasts_imp11_rp_o, contrasts_imp12_rp_o, contrasts_imp13_rp_o, contrasts_imp14_rp_o, contrasts_imp15_rp_o,
  contrasts_imp16_rp_o, contrasts_imp17_rp_o, contrasts_imp18_rp_o, contrasts_imp19_rp_o, contrasts_imp20_rp_o)

```

```{r}
survival_de <- survival_combined %>%
  group_by(group, age) %>%
  summarise(
    survival = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(estimate, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  group_by(group) %>%
  complete(age = 50:66) %>%
  arrange(group, age) %>%
  fill(1:6, .direction = "downup") %>%
  ungroup()

contrast_de_r_h <- contrast_combined_r_h %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_de_r_m <- contrast_combined_r_m %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_de_r_o <- contrast_combined_r_o %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_de_o_m <- contrast_combined_o_m %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

contrast_de_rp_o <- contrast_combined_rp_o %>%
  group_by(age) %>%
  summarise(
    risk_diff = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[3],
    p_value = pool_RR(theta, std.error, conf.level = 0.95, nrow(data %>% filter(age <= age)), 1)[4]) %>%
  mutate(sig = ifelse(p_value < 0.05, "Yes", "No")) %>%
  ungroup()

survival_de_r_h <- survival_de %>% filter(group %in% c("Owners", "Renters")) %>% mutate(country = "Germany") %>% group_by(group, age)
survival_de_r_h <- survival_de_r_h %>% 
  mutate(group = case_when(
    group == "Owners" ~ "Owning",
    group == "Renters" ~ "Renting"))

survival_de_r_m_o <- survival_de %>% filter(group %in% c("Owners outright", "Owners mortgage", "Renters")) %>% mutate(country = "Germany")
survival_de_r_m_o <- survival_de_r_m_o %>% 
  mutate(group = case_when(
    group == "Owners mortgage" ~ "Owning with mortgage",
    group == "Owners outright" ~ "Owning outright",
    group == "Renters" ~ "Renting"))

survival_de_rp_o <- survival_de %>% filter(group %in% c("Owners", "Renters (private)")) %>% mutate(country = "Germany")
survival_de_rp_o <- survival_de_rp_o %>% 
  mutate(group = case_when(
    group == "Owners" ~ "Owning",
    group == "Renters (private)" ~ "Renting (private)"))

contrast_de_r_h <- contrast_de_r_h %>% mutate(country = "Germany", contrast = "Owning")

contrast_de_r_m <- contrast_de_r_m %>% mutate(country = "Germany", contrast = "Owning with mortgage")

contrast_de_r_o <- contrast_de_r_o %>% mutate(country = "Germany", contrast = "Owning outright")

contrast_de_r_m_o <- rbind(contrast_de_r_m, contrast_de_r_o)

contrast_de_o_m <- contrast_de_o_m %>% mutate(country = "Germany", contrast = "Owning outright vs. \nOwning with mortgage")

contrast_de_rp_o <- contrast_de_rp_o %>% mutate(country = "Germany", contrast = "Private renting vs. Owning")

```

```{r}
survival_r_h <- survival_uk_r_h %>% bind_rows(survival_de_r_h)

survival_r_m_o <- survival_uk_r_m_o %>% bind_rows(survival_de_r_m_o)

survival_rp_o <- survival_uk_rp_o %>% bind_rows(survival_de_rp_o)

contrast_r_h <- contrast_uk_r_h %>% bind_rows(contrast_de_r_h)

contrast_r_m <- contrast_uk_r_m %>% bind_rows(contrast_de_r_m)

contrast_r_o <- contrast_uk_r_o %>% bind_rows(contrast_de_r_o)

contrast_r_m_o <- contrast_uk_r_m_o %>% bind_rows(contrast_de_r_m_o)

contrast_o_m <- contrast_uk_o_m %>% bind_rows(contrast_de_o_m)

contrast_rp_o <- contrast_uk_rp_o %>% bind_rows(contrast_de_rp_o)

```

```{r}
plot_survival_r_h <- ggplot(survival_r_h, aes(x = age - 0.5, y = 1 - survival, group = group, fill = group)) +
  facet_grid(~country) +
  geom_stepribbon(aes(ymin = 1 - conf_low, ymax = 1 - conf_high), alpha = 0.25) +
  geom_step(aes(color = group), size = 0.8) +
  scale_x_continuous(NULL, limits = c(50, 66), breaks = seq(50, 65, 1)) +
  scale_y_continuous("Cumulative risk (% retired)", breaks = seq(0, 1, 0.1), label = scales::percent) +
  scale_fill_manual(values = c("darkblue", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkred")) +
  theme_linedraw() +
  coord_cartesian(ylim = c(0,1), xlim = c(51, 65)) +
  theme(legend.position=c(.0825,.91),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        strip.text.x = element_text(size = 13, colour = "white"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank())

plot_survival_r_m_o <- ggplot(survival_r_m_o, aes(x = age - 0.5, y = 1 - survival, group = group, fill = group)) +
  facet_grid(~country) +
  geom_stepribbon(aes(ymin = 1 - conf_low, ymax = 1 - conf_high), alpha = 0.25) +
  geom_step(aes(color = group), size = 0.8) +
  scale_x_continuous(NULL, limits = c(50, 66), breaks = seq(50, 65, 1)) +
  scale_y_continuous("Cumulative risk (% retired)", breaks = seq(0, 1, 0.1), label = scales::percent) +
  scale_fill_manual(values = c("darkblue", "darkgreen", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkgreen", "darkred")) +
  theme_linedraw() +
  coord_cartesian(ylim = c(0,1), xlim = c(51, 65)) +
  theme(legend.position=c(.165,.91),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        strip.text.x = element_text(size = 13, colour = "white"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank())

plot_survival_rp_o <- ggplot(survival_rp_o, aes(x = age - 0.5, y = 1 - survival, group = group, fill = group)) +
  facet_grid(~country) +
  geom_stepribbon(aes(ymin = 1 - conf_low, ymax = 1 - conf_high), alpha = 0.25) +
  geom_step(aes(color = group), size = 0.8) +
  scale_x_continuous(NULL, limits = c(50, 66), breaks = seq(50, 65, 1)) +
  scale_y_continuous("Cumulative risk (% retired)", breaks = seq(0, 1, 0.1), label = scales::percent) +
  scale_fill_manual(values = c("darkblue", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkred")) +
  theme_linedraw() +
  coord_cartesian(ylim = c(0,1), xlim = c(51, 65)) +
  theme(legend.position=c(.165,.91),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        strip.text.x = element_text(size = 13, colour = "white"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank())

plot_contrast_r_h <- ggplot(contrast_r_h, aes(x = age, y = (-1) * risk_diff, group = contrast, shape = contrast)) +
  facet_grid(~country) +
  geom_hline(yintercept = 0, size = 0.5, lty = 2, color = "red") +
  geom_line(color = "lightgrey", position = position_dodge(width = .5)) +
  geom_point(aes(color = sig), size = 2, position = position_dodge(width = .5)) +
  geom_errorbar(aes(ymin = (-1) * conf_low, ymax = (-1) * conf_high, color = sig), width = 0, position = position_dodge(width = .5)) +
  theme_classic() +
  coord_cartesian(xlim = c(51, 65), ylim = c(-0.11, 0.3)) +
  scale_color_manual(values = c("lightgrey", "black")) +
  #scale_color_manual(values = c("black")) +
  guides(colour = "none") +
  theme_linedraw() +
  theme(legend.position = "none",
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        strip.background = element_blank(),
        strip.text.x = element_blank(),
        strip.text.y = element_blank(),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank()) +
  scale_y_continuous("Risk difference (ATE)", breaks = seq(-0.55, 0.55, 0.05), labels = c("-55pp.", "-50pp.", "-45pp.", "-40pp.", "-35pp.", "-30pp.", "-25pp.", "-20pp.", "-15pp.", "-10pp.", "-5pp.", "0pp.", "5pp.", "10pp.", "15pp.", "20pp.", "25pp.", "30pp.", "35pp.", "40pp.", "45pp.", "50pp.", "55pp.")) +
  scale_x_continuous("Age", breaks = seq(50, 70, 1))

plot_contrast_o_m <- ggplot(contrast_o_m, aes(x = age, y = risk_diff, group = contrast, shape = contrast)) +
  facet_grid(~country) +
  geom_hline(yintercept = 0, size = 0.5, lty = 2, color = "red") +
  geom_line(color = "lightgrey", position = position_dodge(width = .5)) +
  geom_point(aes(color = sig), size = 2, position = position_dodge(width = .5)) +
  geom_errorbar(aes(ymin = conf_low, ymax = conf_high, color = sig), width = 0, position = position_dodge(width = .5)) +
  theme_classic() +
  coord_cartesian(xlim = c(51, 65), ylim = c(-0.11, 0.3)) +
  scale_color_manual(values = c("lightgrey", "black")) +
  #scale_color_manual(values = c("black")) +
  guides(colour = "none") +
  theme_linedraw() +
  theme(legend.position = c(.175,.90),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        strip.background = element_blank(),
        strip.text.x = element_blank(),
        strip.text.y = element_blank(),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank()) +
  scale_y_continuous("Risk difference (ATE)", breaks = seq(-0.55, 0.55, 0.05), labels = c("-55pp.", "-50pp.", "-45pp.", "-40pp.", "-35pp.", "-30pp.", "-25pp.", "-20pp.", "-15pp.", "-10pp.", "-5pp.", "0pp.", "5pp.", "10pp.", "15pp.", "20pp.", "25pp.", "30pp.", "35pp.", "40pp.", "45pp.", "50pp.", "55pp.")) +
  scale_x_continuous("Age", breaks = seq(50, 70, 1))

plot_contrast_r_m_o <- ggplot(contrast_r_m_o, aes(x = age, y = (-1) * risk_diff, group = contrast, shape = contrast)) +
  facet_grid(~country) +
  geom_hline(yintercept = 0, size = 0.5, lty = 2, color = "red") +
  geom_line(color = "lightgrey", position = position_dodge(width = .5)) +
  geom_point(aes(color = sig), size = 2, position = position_dodge(width = .5)) +
  geom_errorbar(aes(ymin = (-1) * conf_low, ymax = (-1) * conf_high, color = sig), width = 0, position = position_dodge(width = .5)) +
  theme_classic() +
  coord_cartesian(xlim = c(51, 65), ylim = c(-0.11, 0.3)) +
  scale_color_manual(values = c("lightgrey", "black")) +
  #scale_color_manual(values = c("black")) +
  guides(colour = "none") +
  theme_linedraw() +
  theme(legend.position = c(.175,.90),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        strip.background = element_blank(),
        strip.text.x = element_blank(),
        strip.text.y = element_blank(),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank()) +
  scale_y_continuous("Risk difference (ATE)", breaks = seq(-0.55, 0.55, 0.05), labels = c("-55pp.", "-50pp.", "-45pp.", "-40pp.", "-35pp.", "-30pp.", "-25pp.", "-20pp.", "-15pp.", "-10pp.", "-5pp.", "0pp.", "5pp.", "10pp.", "15pp.", "20pp.", "25pp.", "30pp.", "35pp.", "40pp.", "45pp.", "50pp.", "55pp.")) +
  scale_x_continuous("Age", breaks = seq(50, 70, 1))

plot_contrast_rp_o <- ggplot(contrast_rp_o, aes(x = age, y = -risk_diff, group = contrast, shape = contrast)) +
  facet_grid(~country) +
  geom_hline(yintercept = 0, size = 0.5, lty = 2, color = "red") +
  geom_line(color = "lightgrey", position = position_dodge(width = .5)) +
  geom_point(aes(color = sig), size = 2, position = position_dodge(width = .5)) +
  geom_errorbar(aes(ymin = -conf_low, ymax = -conf_high, color = sig), width = 0, position = position_dodge(width = .5)) +
  theme_classic() +
  coord_cartesian(xlim = c(51, 65), ylim = c(-0.11, 0.3)) +
  scale_color_manual(values = c("lightgrey", "black")) +
  scale_color_manual(values = c("lightgrey", "black")) +
  guides(colour = "none") +
  theme_linedraw() +
  theme(legend.position = c(.19,.90),
        text = element_text(colour = "black"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.title.x = element_text(colour = "black", face = "bold", size = 13),
        axis.title.y = element_text(colour = "black", face = "bold", size = 13),
        axis.line = element_line(),
        axis.ticks = element_line(),
        axis.text.x = element_text(size = 11, color = "black"),
        axis.text.y = element_text(size = 11, color = "black"),
        strip.background = element_blank(),
        strip.text.x = element_blank(),
        strip.text.y = element_blank(),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        strip.placement = 'outside',
        panel.background = element_rect(fill = "transparent", colour = NA_character_),
        plot.background = element_rect(fill = "transparent", colour = NA_character_),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.text = element_text(colour = "black", size = 12),
        legend.title = element_blank()) +
  scale_y_continuous("Risk difference (ATE)", breaks = seq(-0.55, 0.55, 0.05), labels = c("-55pp.", "-50pp.", "-45pp.", "-40pp.", "-35pp.", "-30pp.", "-25pp.", "-20pp.", "-15pp.", "-10pp.", "-5pp.", "0pp.", "5pp.", "10pp.", "15pp.", "20pp.", "25pp.", "30pp.", "35pp.", "40pp.", "45pp.", "50pp.", "55pp.")) +
  scale_x_continuous("Age", breaks = seq(50, 70, 1))

```

```{r}
plot_1 <- grid.arrange(plot_survival_r_h, plot_contrast_r_h, nrow = 2)
ggsave(".../plot_results_1_975.pdf", height = 10.5, width = 7.5, bg = "transparent", plot_1)

plot_2 <- grid.arrange(plot_survival_r_m_o, plot_contrast_r_m_o, nrow = 2)
ggsave(".../plot_results_2_975.pdf", height = 10.5, width = 7.5, bg = "transparent", plot_2)

plot_3 <- grid.arrange(plot_survival_r_m_o, plot_contrast_o_m, nrow = 2)
ggsave(".../plot_results_3_975.pdf", height = 10.5, width = 7.5, bg = "transparent", plot_3)

plot_4 <- grid.arrange(plot_survival_rp_o, plot_contrast_rp_o, nrow = 2)
ggsave(".../plot_results_4_975.pdf", height = 10.5, width = 7.5, bg = "transparent", plot_4)

```
